Youtube downloader
This is a simple script to download youtube videos with high or low resolution or audio format, it is a simple project mainly to learn the basics, the next step is to make it a GUI app. feel free to share ideas about things to add or resources to learn from. Thank you.